requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-injector==0.15.0
boto3==1.34.108
APScheduler==3.10.4
python-dotenv==1.0.1
schema==0.7.7
pytest==8.2.0
pytest-cov==5.0.0
retry==0.9.2